JoinFormPageForActiveOrNewUser → JoinDispatcher 으로 정리
- 호출 측은 /join/form 한 경로만 알도록 유지하면서, 신규/재가입 외에 onboarding 미완·대기열 케이스로 분기를 확장하려면 dispatcher 가 4갈래를 알아야 함
- 신규 사용자 분기는 OnboardingPage 가 대체하므로 JoinFormPageForNewUser / JoinFormCardForNewUser 와 그 전용 타입은 죽은 코드
- 라우터의 join/form/new-user 경로를 제거하고 join/onboarding · join/complete 를 정식 경로로 노출해 RootRedirect / dispatcher 의 navigate 대상을 단일화
