+++ apps/web/src/login/model/join.ts
@@ -1,10 +1,3 @@
-export interface JoinFormDataForNewUser {
-  name: string
-  phoneNumber: string
-  nickname?: string
-  referrer: string
-}
-
 export interface JoinFormDataForActiveUser {
   keep?: string
   problem?: string
